server/requirements.txt
aiofiles==23.1.0
anyio==3.6.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==4.4.2
deprecation==2.1.0
docx2pdf==0.1.8
fastapi==0.93.0
fire==0.5.0
fonttools==4.39.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
img2pdf==0.4.4
lxml==4.9.2
moviepy==1.0.3
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pdf2docx==0.5.6
pdf2pptx==1.0.5
pikepdf==7.1.1
Pillow==9.4.0
proglog==0.1.10
pydantic==1.10.6
PyMuPDF==1.20.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pywin32==306
PyYAML==6.0
requests==2.30.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.25.0
termcolor==2.2.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
XlsxWriter==3.0.9
